Conflicts:
	Source/Core/Core/HW/GCPadEmu.cpp
	Source/Core/Core/HW/WiimoteEmu/WiimoteEmu.cpp
	Source/Core/InputCommon/InputCommon.vcxproj.filters

+++ Source/Core/InputCommon/ControllerInterface/Device.h
@@ -10,7 +10,7 @@
 #include "Common/Common.h"
 
 // idk in case I wanted to change it to double or something, idk what's best
-typedef float ControlState;
+typedef double ControlState;
 
 namespace ciface
 {
